PHPCS: add rule for "declare symbols or run code, but not both".
This was disabled in project-template as being too opinionated, but it is often
a helpful rule to have enabled, so this commit turns it on.

As a consequence of removing the "block file being run directly" checks, we
should assume that a well-configured web server will only allow access to
WP's `index.php` and `admin-ajax.php`, etc.